# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PillMind API",
    "description": "PillMind 3.0 Kurumsal Entegrasyon, HL7 FHIR Uyumlu ve Güvenli İlaç Etkileşim Analiz API Sözleşmesi",
    "version": "3.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Lokal Geliştirme Sunucusu"
    }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "Kullanıcı Kaydı",
        "description": "E-posta tabanlı şifresiz kayıt oluşturur.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "doktor@pillmind.com"
                  }
                },
                "required": ["email"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Başarılı kayıt ve otomatik giriş işlemi. Çerez (HttpOnly session) atanır.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "email": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Geçersiz e-posta formatı veya eksik parametre" }
        }
      }
    },
...
    "/api/check": {
      "post": {
        "summary": "Deterministik İlaç Etkileşimi Sorgulama",
        "description": "Verilen ilaç ID listesi için veri tabanındaki veya yerel JSON fallback üzerindeki etkileşimleri, dozaşımı uyarılarını, gıda etkileşimlerini ve kontrendikasyonları hesaplar.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "drugIds": {
                    "type": "array",
                    "items": { "type": "string" },
                    "example": ["aspirin", "warfarin"]
                  },
                  "patientContext": {
                    "type": "object",
                    "properties": {
                      "isPregnant": { "type": "boolean", "example": false },
                      "isBreastfeeding": { "type": "boolean", "example": false },
                      "ageGroup": { "type": "string", "example": "adult" },
                      "renalRisk": { "type": "boolean", "example": false },
                      "hepaticRisk": { "type": "boolean", "example": false },
                      "diseases": {
                        "type": "array",
                        "items": { "type": "string" },
                        "example": ["K25"]
                      }
                    }
                  }
                },
                "required": ["drugIds"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Başarılı etkileşim analizi.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "interactions": { "type": "array", "items": { "type": "object" } },
                    "accumulationWarnings": { "type": "array", "items": { "type": "object" } },
                    "foodInteractions": { "type": "array", "items": { "type": "object" } },
                    "contraindications": { "type": "array", "items": { "type": "object" } },
                    "polypharmacyReport": { "type": "object" },
                    "checkedDrugs": { "type": "array", "items": { "type": "string" } },
                    "totalFound": { "type": "integer" },
                    "disclaimer": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "Eksik veya geçersiz parametre formatı" },
          "429": { "description": "Çok fazla istek (IP tabanlı rate-limiting)" }
        }
      }
    },
...
  }
}